+++ src/funding/paypal/config.jsx
@@ -34,6 +34,8 @@ export function getPayPalConfig(): FundingSourceConfig {
       BUTTON_COLOR.SILVER,
       BUTTON_COLOR.BLACK,
       BUTTON_COLOR.WHITE,
+      BUTTON_COLOR.DEFAULT_BLUE_LIGHT_BLUE,
+      BUTTON_COLOR.DEFAULT_BLUE_DARK_BLUE,
     ],
 
     logoColors: {
@@ -42,6 +44,8 @@ export function getPayPalConfig(): FundingSourceConfig {
       [BUTTON_COLOR.BLUE]: LOGO_COLOR.WHITE,
       [BUTTON_COLOR.BLACK]: LOGO_COLOR.WHITE,
       [BUTTON_COLOR.WHITE]: LOGO_COLOR.BLUE,
+      [BUTTON_COLOR.DEFAULT_BLUE_LIGHT_BLUE]: LOGO_COLOR.BLACK,
+      [BUTTON_COLOR.DEFAULT_BLUE_DARK_BLUE]: LOGO_COLOR.BLUE,
     },
 
     labelText: ({ content, label, period }) => {