+++ src/funding/paypal/template.jsx
@@ -11,7 +11,9 @@ import {
   PPLogoExternalImage,
   PPLogoInlineSVG,
   PayPalLogoExternalImage,
+  PayPalRebrandLogoExternalImage,
   PayPalLogoInlineSVG,
+  PayPalRebrandLogoInlineSVG,
   CreditLogoExternalImage,
   CreditLogoInlineSVG,
   CreditMarkExternalImage,
@@ -46,11 +48,29 @@ import {
 
 import css from "./style.scoped.scss";
 
-export function Logo({ logoColor }: LogoOptions): ChildType {
+export function Logo({ logoColor, experiment }: LogoOptions): ChildType {
+  const { isPaypalRebrandEnabled, defaultBlueButtonColor } = experiment;
+
+  if (defaultBlueButtonColor === "gold" || !isPaypalRebrandEnabled) {
+    // csnw globals.js
+    return __WEB__ ? (
+      // helps reduce bundle size by fetching logos
+      <PayPalLogoExternalImage logoColor={logoColor} />
+    ) : (
+      // cdnx/sdk-logo/xxxx/paypal-gold.svg
+      // use for server side rendering
+      <PayPalLogoInlineSVG logoColor={logoColor} />
+    );
+  }
+
+  // csnw globals.js
   return __WEB__ ? (
-    <PayPalLogoExternalImage logoColor={logoColor} />
+    // helps reduce bundle size by fetching logos
+    <PayPalRebrandLogoExternalImage logoColor={logoColor} />
   ) : (
-    <PayPalLogoInlineSVG logoColor={logoColor} />
+    // cdnx/sdk-logo/xxxx/paypal-gold.svg
+    // use for server side rendering
+    <PayPalRebrandLogoInlineSVG logoColor={logoColor} />
   );
 }