+++ src/ui/buttons/buttons.jsx
@@ -251,6 +251,7 @@ export function Buttons(props: ButtonsProps): ElementNode {
       ].join(" ")}
     >
       <Style
+        experiment={experiment}
         nonce={nonce}
         style={style}
         fundingEligibility={fundingEligibility}