+++ src/ui/buttons/style.jsx
@@ -4,6 +4,8 @@
 import { node, type ElementNode } from "@krakenjs/jsx-pragmatic/src";
 import { type FundingEligibilityType } from "@paypal/sdk-constants/src";
 
+import type { Experiment } from "../../types";
+
 import { type ButtonStyle } from "./props";
 import { componentStyle } from "./styles";
 
@@ -11,12 +13,14 @@ type StyleProps = {|
   style: ButtonStyle,
   nonce: string,
   fundingEligibility: FundingEligibilityType,
+  experiment: Experiment,
 |};
 
 export function Style({
   style,
   nonce,
   fundingEligibility,
+  experiment,
 }: StyleProps): ElementNode {
   const { height, disableMaxWidth, disableMaxHeight, borderRadius } = style;
   const css = componentStyle({
@@ -25,6 +29,7 @@ export function Style({
     disableMaxWidth,
     disableMaxHeight,
     borderRadius,
+    experiment,
   });
 
   return <style nonce={nonce} innerHTML={css} />;